+++ include/dxc/CMakeLists.txt
@@ -17,7 +17,6 @@ configure_file(
 
 add_subdirectory(DXIL)
 add_subdirectory(DxilContainer)
-add_subdirectory(HLSL)
 add_subdirectory(Support)
 add_subdirectory(Tracing)
