--- include/dxc/HLSL/CMakeLists.txt
@@ -1,1 +1,0 @@
-add_hlsl_hctgen(DxilValidationInc OUTPUT DxilValidation.inc BUILD_DIR)
