+++ lib/CMakeLists.txt
@@ -29,6 +29,7 @@ add_subdirectory(DxilPdbInfo) # HLSL Change
 add_subdirectory(DxilPIXPasses) # HLSL Change
 add_subdirectory(DxilDia) # HLSL Change
 add_subdirectory(DxilRootSignature) # HLSL Change
+add_subdirectory(DxilValidation) # HLSL Change
 add_subdirectory(DxcBindingTable) # HLSL Change
 add_subdirectory(DxrFallback) # HLSL Change
 add_subdirectory(DxilCompression) # HLSL Change
