+++ tools/clang/tools/dxcompiler/dxclinker.cpp
@@ -23,8 +23,8 @@
 #include "llvm/ADT/SmallVector.h"
 #include <algorithm>
 
+#include "dxc/DxilValidation/DxilValidation.h"
 #include "dxc/HLSL/DxilLinker.h"
-#include "dxc/HLSL/DxilValidation.h"
 #include "dxc/Support/HLSLOptions.h"
 #include "dxc/Support/Unicode.h"
 #include "dxc/Support/microcom.h"
@@ -37,6 +37,7 @@
 #include "llvm/IR/LLVMContext.h"
 #include "llvm/IR/Module.h"
 #include "llvm/Support/raw_ostream.h"
+
 
 using namespace hlsl;
 using namespace llvm;
