+++ tools/clang/tools/dxcvalidator/CMakeLists.txt
@@ -7,7 +7,7 @@ set( LLVM_LINK_COMPONENTS
   dxcsupport
   DXIL
   DxilContainer
-  HLSL
+  DxilValidation
   Option     # option library
   Support    # just for assert and raw streams
   )
